// src/api/model/featureConfig.ts
/**
 * Receipt Wrangler API.
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface FeatureConfig { 
    /**
     * Whether AI powered receipts are enabled
     */
    aiPoweredReceipts: boolean;
    /**
     * Whether local sign up is enabled
     */
    enableLocalSignUp: boolean;
}